HSEARCH-2065 Centralize the FacetComparators
